### 2025-12-15 #190 @editorialbot
[PRE REVIEW]: GraphLab.jl: A Julia Framework for Graph Partitioning
**Submitting author:** @lechekhabm (<a href="http://orcid.org/0009-0008-2279-0804">Malik Lechekhab</a>)
**Repository:** https://github.com/lechekhabm/GraphLab.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
**Managing EiC:** Dimitris Floros

Thanks for submitting your paper to JCON @lechekhabm. **Currently, there isn't a JCON editor assigned** to your paper.
