### 2025-12-20 #190 @fcdimitr
Hi @lechekhabm, and thank you for submitting your manuscript to the JuliaCon proceedings. I’ll be serving as the editor for this submission.

I will begin identifying reviewers for this prereview issue and will initiate the formal review process in a separate GitHub issue once everything is ready.
In the meantime, do you have any suggestions for potential reviewers? If so, please share a list of GitHub handles.

Also, I noticed a few missing or invalid DOIs. Could you please take a look at those? (There may be some false negatives.)

Thank you.
