### 2026-02-13 05e9285
Add status badge for JuliaCon presentation
