### 2026-02-19 #190 @fcdimitr
Hi @mipals 👋 ,

would you like to review this JuliaCon proceedings submission?

you can read more about reviewing for JuliaCon proceedings in the [reviewer guide](https://juliacon.github.io/proceedings-guide/reviewer/).
