### 2026-02-25 #190 @andreasw987
Hi @fcdimitr ,
I'm not familar with this process.  Reading the reviewer guide, it seems that there should be a checklist for me:

[Checklist](https://juliacon.github.io/proceedings-guide/reviewer/#checklist)

"A reviewer checklist is generated for each reviewer of the paper. A box should be checked only once the reviewer considers that the criterion is met by the work presented."

Where is that?
