### 2026-02-26 #196 @editorialbot
[REVIEW]: GraphLab.jl: A Julia Framework for Graph Partitioning
**Submitting author:** @lechekhabm (<a href="http://orcid.org/0009-0008-2279-0804">Malik Lechekhab</a>)
**Repository:** https://github.com/lechekhabm/GraphLab.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
