### 2026-02-27 lechekhabm/GraphLab.jl#6 @mipals
Suggestions
Hi there,

I am currently reviewing this package JuliaCon/proceedings-review#196. While I do not see anything blocking the review, I have a few two suggestions for improvements (optional and up to you if you think its worth the effort).

- The two examples could be exported using [Literate.jl](https://github.com/fredrikekre/Literate.jl) and included directly in the documentation.
- The package currently adds quite a few png images of high resolution causing the size of the package to be unnecessarily high. What makes it slightly worse is that the images are both in the examples folder and paper folder.

Cheers,
Mikkel
