### 2026-03-02 #196 @mipals
I am finished looking through the paper, the code, and the documentation. Except for a few missing DOIs I think everything looks good.
