### 2026-03-05 #196 @andreasw987
I finished my review. I have a few comments; except for 3 of them, they are suggestions on how to improve the description of the algorithms.
