### 2026-04-01 08dc6cd
Update image path for Coordinate Bisection
