### 2026-04-01 4ca9543
Update image paths in README.md
