### 2026-04-02 7cdd1be
Add Literate.jl examples to documentation
