### 2026-04-02 9f434a3
Ensure Literate examples print tables in docs
