+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API-Reference","page":"API Reference","title":"API Reference","text":"","category":"section"},{"location":"api/","page":"API Reference","title":"API Reference","text":"This section lists the public functions and types exported by GraphLab.jl.","category":"page"},{"location":"api/","page":"API Reference","title":"API Reference","text":"Modules = [GraphLab]\nPrivate = false","category":"page"},{"location":"api/#GraphLab.build_adjacency-Tuple{Matrix{Int64}, Int64}","page":"API Reference","title":"GraphLab.build_adjacency","text":"build_adjacency(edges::Matrix{Int}, num_nodes::Int)\n\nConstruct the adjacency matrix of a graph from an edge list.\n\nArguments\n\nedges::Matrix: Matrix where each row represents an edge [u, v].\nnum_nodes::Int: Total number of nodes in the graph.\n\nReturns\n\nA symmetric sparse adjacency matrix (SparseMatrixCSC{Int, Int}).\n\nExample\n\njulia> edges = [1 2; 2 3; 3 1]\njulia> A = build_adjacency(edges, 3)\n\n\n\n\n\n","category":"method"},{"location":"api/#GraphLab.build_adjacency-Tuple{String}","page":"API Reference","title":"GraphLab.build_adjacency","text":"build_adjacency(type::String)\n\nGenerate a predefined adjacency matrix and corresponding node coordinates.\n\nArguments\n\ntype::String: Type of graph to generate.  \n\"network\": A predefined network structure.  \n\"triangles\": A small triangular mesh structure.\n\nReturns\n\nA::SparseMatrixCSC: The sparse adjacency matrix of the graph.\ncoords::Matrix: Node coordinates for visualization.\n\nExample\n\njulia> A, coords = build_adjacency(\"network\")\n\n\n\n\n\n","category":"method"},{"location":"api/#GraphLab.compute_partition_balance-Tuple{AbstractVector}","page":"API Reference","title":"GraphLab.compute_partition_balance","text":"compute_partition_balance(p::AbstractVector) -> Float64\n\nComputes the balance metric of a given graph partitioning.\n\nParameters\n\np::AbstractVector: A vector where p[i] represents the partition index assigned to vertex i.\n\nReturns\n\nFloat64: The balance metric, defined as the ratio of the largest partition size to the ideal partition size. A value close to 1.0 indicates a well-balanced partitioning, while higher values suggest imbalance.\n\nExample\n\njulia> p = [1, 1, 2, 2, 2, 3, 3, 3, 3]  # Example partitioning\njulia> balance = compute_partition_balance(p)\njulia> println(balance)  # Output close to 1 for balanced partitions\n\n\n\n\n\n","category":"method"},{"location":"api/#GraphLab.count_edge_cut-Tuple{AbstractMatrix, AbstractVector}","page":"API Reference","title":"GraphLab.count_edge_cut","text":"count_edge_cut(A::AbstractMatrix, p::AbstractVector)\n\nCount the number of edges that cross partitions in the graph A.\n\nArguments\n\nA::AbstractMatrix: Adjacency matrix of the graph.\np::AbstractVector: Partition vector where p[v] represents the partition of vertex v.\n\nReturns\n\nThe number of edges that connect nodes in different partitions.\n\nExample\n\njulia> count_edge_cut(A, p)\n15\n\n\n\n\n\n","category":"method"},{"location":"api/#GraphLab.draw_graph-Tuple{SparseArrays.SparseMatrixCSC, Matrix, Vector{Int64}}","page":"API Reference","title":"GraphLab.draw_gra
[…]
