### 2026-04-03 d1e9ee8
Ensure Literate examples print tables in docs
