+++ src/tools/part_randsphere.jl
@@ -16,18 +16,18 @@ coords: n × d matrix of points
 Returns: n × (d+1) matrix on the unit sphere
 """
 function _stereoup(coords::Matrix{Float64})
-    n, d = size(coords)
-    lifted = zeros(n, d+1)
+  n, d = size(coords)
+  lifted = zeros(n, d + 1)
 
-    for i in 1:n
-        x = coords[i, :]
-        r2 = dot(x, x)
-        lifted[i, 1:d] .= 2 * x
-        lifted[i, d+1] = r2 - 1
-        lifted[i, :] ./= r2 + 1
-    end
+  for i in 1:n
+    x = coords[i, :]
+    r2 = dot(x, x)
+    lifted[i, 1:d] .= 2 * x
+    lifted[i, d+1] = r2 - 1
+    lifted[i, :] ./= r2 + 1
+  end
 
-    return lifted
+  return lifted
 end
 
 
@@ -42,11 +42,11 @@ Returns:
 - center::Vector{Float64}: estimated _centerpoint
 """
 function _centerpoint(X::Matrix{Float64}, sample_size::Int)
-    n, d = size(X)
-    idx = rand(1:n, min(sample_size, n))
-    sample = X[idx, :]
-    c = mapslices(median, sample; dims=1)[:]
-    return c
+  n, d = size(X)
+  idx = rand(1:n, min(sample_size, n))
+  sample = X[idx, :]
+  c = mapslices(median, sample; dims=1)[:]
+  return c
 end
 
 
@@ -62,14 +62,14 @@ end
 # """
 # function _conmap(c::Vector{Float64}, xyz::Matrix{Float64})
 #     d = size(xyz, 2)
-  
+
 #     # Compute reflection and stretch
 #     Q, r = _reflector(c)
 #     alpha = sqrt((1+r)/(1-r))
-  
+
 #     # Reflect
 #     xyzref = xyz * Q;
-  
+
 #     # Handle north pole in the stereographic projection
 #     norths = findall(x -> abs(x - 1) < sqrt(eps()), xyzref[:, d]) 
 #     if !isempty(norths)
@@ -77,54 +77,54 @@ end
 #       # xyzref[norths, :] = zeros(length(norths), d)
 #       xyzref[norths, :] .= 0
 #     end
-  
+
 #     xyref = _stereodown(xyzref)
-  
+
 #     # Stretch
 #     xymap = xyref ./ alpha
 #     xyzmap = _stereoup(xymap)
-  
+
 #     if !isempty(norths)
 #       xyzmap[norths,:] = xyzn
 #     end
 
 #     println(xyzmap)
-  
+
 #     return xyzmap, xymap
 # end
 
 function _conmap(c::Vector{Float64}, xyz::Matrix{Float64})
-    d = size(xyz, 2)
+  d = size(xyz, 2)
 
-    # 1. reflection / rotation
-    Q, r = _reflector(c)
-    alpha = sqrt((1 + r) / (1 - r))   # stretch factor (as in the original code)
+  # 1. reflection / rotation
+  Q, r = _reflector(c)
+  alpha = sqrt((1 + r) / (1 - r))   # stretch factor (as in the original code)
 
-    # 2. rotate all points
-    xyzref = xyz * Q
+  # 2. rotate all points
+  xyzref = xyz * Q
 
-    # 3. handle north pole (z == 1) specially to avoid stereographic blowup
-    norths = findall(x -> abs(x - 1) < sqrt(eps(Float64)), xyzref[:, d])
-    xyzn = xyzref[norths, :]
-    if !isempty(norths)
-        xyzref[norths, :] .= 0
-    end
+  # 3. handle north pole (z == 1) specially to avoid stereographic blowup
[…]
