### 2026-04-29 d58061e
Add Literate tutorials for package examples
