### 2026-04-29 11cb8de
Add CairoMakie to docs environment
