### 2026-05-02 f7695b3
Prepare JuliaCon proceedings submission
