+++ paper/paper.tex
@@ -15,6 +15,7 @@
 \usepackage{subcaption}
 
 % refs
+\usepackage{hyperref}
 \usepackage{cleveref}
 \usepackage{placeins}
 
@@ -24,6 +25,9 @@
   breakatwhitespace=true,
   columns=fullflexible
 }
+
+% MAY BREAK THINGS, DOUBLE CHECK THROUGH THE PAPER NEEDED
+%\usepackage[strings]{underscore}
 
 % Define the \add command
  \newcommand{\add}[1]{\textit{\textcolor{blue}{#1}}}
