+++ paper/ref.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={65--98},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/141000671}
 }
 
 @article{Moore1990,
@@ -146,12 +147,16 @@ doi = {10.1137/S1064827594275339},
   pages={418-427},
   doi={10.1109/IPPS.1995.395965}}
 
-
-@article{sasidharan2015space,
-  title={Space-filling curves for partitioning adaptively refined meshes},
-  author={Sasidharan, Aparna and Snir, Marc},
-  journal={Mathematics and Computer Science},
-  year={2015}
+@techreport{SasidharanSnir2015SFC,
+  author      = {Aparna Sasidharan and Marc Snir},
+  title       = {Space-filling Curves for Partitioning Adaptively Refined Meshes},
+  institution = {Argonne National Laboratory, Mathematics and Computer Science Division},
+  type        = {Technical Report},
+  number      = {P5355-0615},
+  address     = {Argonne, IL, USA},
+  xxxnote        = {, 2015},
+  year        = {2015},
+  url         = {https://www.mcs.anl.gov/papers/P5355-0615.pdf}
 }
 
 @INPROCEEDINGS{Sasidharan15,
@@ -187,7 +192,8 @@ Xeprint = {
   volume={290},
   year={1976},
   publisher={Macmillan London},
-  pages = {74-75}
+  pages = {74-75},
+  doi = {10.1137/1021086}
 }
 
 
@@ -232,7 +238,8 @@ keywords = {Graph partitioning, hypergraph partitioning, load balancing}
   volume    = {2},
   number    = {2},
   pages     = {135--148},
-  year      = {1991}
+  year      = {1991},
+  doi       = {10.1016/0956-0521(91)90014-v}
 }
 
 
@@ -254,7 +261,8 @@ keywords = {Graph partitioning, hypergraph partitioning, load balancing}
   volume    = {25},
   number    = {4},
   pages     = {619--633},
-  year      = {1975}
+  year      = {1975},
+  doi       = {10.21136/CMJ.1975.101357}
 }
 
 @article{Simon91,
@@ -379,17 +387,6 @@ doi = {10.1137/S1064827593255135},
     abstract = { The most commonly used p-way partitioning method is recursive bisection (RB). It first divides a graph or a mesh into two equal-sized pieces, by a "good" bisection algorithm, and then recursively divides the two pieces. Ideally, we would like to use an optimal bisection algorithm. Because the optimal bisection problem that partitions a graph into two equal-sized subgraphs to minimize the number of edges cut is NP-complete, practical RB algorithms use more efficient heuristics in place of an optimal bisection algorithm. Most such heuristics are designed to find the best possible bisection within allowed time. We show that the RB method, even when an optimal bisection algorithm is assumed, may produce a p-way partition that is very far way from the optimal one. Our negative result is complemented by two positive ones: first we show that for some important classes of graphs that occur in practical applications, such as well-shaped finite-element and finite-difference meshes, RB is within a constant factor of the optimal one "almost always." Second, we show that if the balance condition is relaxed so that each block in the p-way partition is bounded by 2n/p, where n is the number of vertices of the graph, then a modified RB finds an approximately balanced p-way partition whose cost is within an O(log p) factor of the cost of the optimal p-way partition. }
 }
  
- 
-@misc{Lech25,
-      title={GraphLab.jl: A Julia Framework for Graph Partitioning}, 
-      author={Malik Lechekhab and Dimosthenis Pasadakis and Aryan Eftekhari and Roger Käppeli and Olaf Schenk},
-      year={2025},
[…]
