### 2026-05-02 lechekhabm/GraphLab.jl#6 @lechekhabm
Thank you @mipals for these suggestions, which improved the exposition of the package's documentation.

> The two examples could be exported using [Literate.jl](https://github.com/fredrikekre/Literate.jl) and included directly in the documentation.

We have incorporated Literate.jl into the documentation workflow. Rather than exporting the two original example scripts directly, which resulted in long and difficult-to-read code blocks, we reorganized the material into smaller tutorial-style Literate examples. These examples now present the main workflows, including a quick start, a comparison of bisection methods, and recursive partitioning. The full example scripts are kept in the \texttt{examples/} directory.

> The package currently adds quite a few png images of high resolution causing the size of the package to be unnecessarily high. What makes it slightly worse is that the images are both in the examples folder and paper folder.

We have removed the image folder from the package, thereby reducing its size and eliminating this unnecessary duplication. The high-resolution PNG images are now present only in the paper/folder .
