### 2026-05-03 #196 @lechekhabm
Dear @fcdimitr, @andreasw987, and @mipals,

We have pushed a revised version of the paper addressing your comments, and the updated article proof has now been generated by @editorialbot .

We have also answered the comments in the corresponding issues:

- [Mikkel's comments](https://github.com/lechekhabm/GraphLab.jl/issues/6#issuecomment-4363925585)
- [Andreas' comments](https://github.com/lechekhabm/GraphLab.jl/issues/7#issuecomment-4364008505)

Please let us know if anything remains unclear or if further changes are needed.

Many thanks again for your review.
