### 2026-05-03 d7e4ff9
Prepare JuliaCon proceedings submission - minor edits
