### 2026-06-17 4850d6f
Fix badge duplication in README.md
