### 2026-07-01 #196 @editorialbot
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
